+++ src/collective/messagesviewlet/subscribers.py
@@ -6,5 +6,5 @@ def change_hidden_uid(message, event):
     """
         Generate a new uid if the message is deactivated
     """
-    if event.action == 'disactivate':
+    if event.action == 'deactivate':
         message.hidden_uid = generate_uid()